TVB-2761: export linked DT's on a separate button keeping the old dt single file export
