+++ framework_tvb/tvb/interfaces/web/templates/jinja2/project/details_datatype_overlay.html
@@ -68,7 +68,7 @@
             <ul>
                 {% for exporter in exporters %}
                     <li>
-                        <button class="action action-export" onclick="window.location='/project/downloaddata/{{ entity_gid }}/{{ exporter }}';">{{ exporters[exporter] }}</button>
+                        <button class="action action-export" onclick="window.location='/project/downloaddatatype/{{ entity_gid }}/{{ exporter }}';">{{ exporters[exporter] }}</button>
                     </li>
                 {% endfor %}
             </ul>
